Improve the scripts - mention that ps3 keys are required to use some of the tools

+++ dev_flash_extractor_osx.sh
@@ -3,6 +3,13 @@ echo "dev_flash & dev_flash3 extractor (For OSX)"
 echo ""
 
 if [ $# -eq 1 ]; then
+
+    # Need the PS3 keys otherwise next steps won't work
+    if [ ! -d .ps3 ];then
+      echo "PS3 folder '.ps3' doesn't exist!, please copy the requires ps3 keys to that folder"
+      exit 1
+    fi
+
 	echo "Extracting PUP.."
 	$TOOLS/pupunpack $1 TMP >> logs.txt
 	rm -rf logs.txt
